Image Population Date Range Handling
This commit fixes the Android GPS tracking service when the device
screen is flipped, as well as handles the center on me button click
event.
This commit handles the date event slider to populate sample field
images within the map view. It is a crude method based on already stored
information that would want to be dynamically allocated later. But the
principles are there.