Stuff Changed and Adding Bluetooth Service
The purpose of this upload is to begin supporting a bluetooth remote
connection service. The importance and initial thought of this is to
enable two devices to communicate to eachother. The thought process
behind this was because each pilot would be flying while another
operator (primarily civil) would need to point/aim the camera at objects
of interest. To accomplish this over serial only on device connection to
the aircraft is committed. Therefore, the secondary device would talk
via bluetooth to primary which could then parse and transmit the
commands to the aircraft.