A Few Additional Icons
The thought is to have several different icon options possible when
planning a mission. This can be seen as dropped on a map or currently
selected and dragging or editing. The same mentality should be used on
the map representing the active target waypoint and other waypoints in
the mission.