Removed Images to Load Faster
The problem with static loading the images resulted in long load times
to deploy app to device. There is a working dynamic load that will be
vetted out in the next few days. Until then, I will duplicate the one
field onto other indices to force the panel side bar to populate if
needed.