Base Class Parameter Inheritance Beginning
The purpose of this was to set up an easy method of assigning the
parameters available to each object that could then be transmitted to
the vehicle. Also, added a flag for whether or not WP should be sent to
vehicle so that safety WP could be added.

Lastly, started migrating completely to the WPHandler which is managing
all WP and events.